+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.3.3
 - Add ModMenu Support (Fabric) 
 - Fix Oddion Despawning 
+- Give ManOWar Custom Persistence
 
 # 2.3.2
 - Remove Milk from Allium Oddion Soup Recipe, Change Nutrition value from 9 to 8